Move pipeline definitions to the top-level of the repo
And split them into multiple files, for readability.
