+++ .gitignore
@@ -1,5 +1,4 @@
 *.tt
-*.yaml
 
 .idea/
 
@@ -19,3 +18,6 @@ workdir/mturk-batches
 # IDEs
 .idea/
 .vscode/
+
+.ipynb_checkpoints
+__pycache__
