+++ LICENSE
@@ -1,5 +1,5 @@
 The MIT License (MIT)
-Copyright (c) 2019 The Board of Trustees of the Leland Stanford Junior University
+Copyright (c) 2019-2020 The Board of Trustees of the Leland Stanford Junior University
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
