+++ components/auto-annotate-selftrain.yaml
