+++ components/sts-paraphrase.yaml
