+++ components/train-trade.yaml
