+++ config.sample
@@ -1,15 +1,14 @@
+# S3 bucket to use
 S3_BUCKET=
+# profile to use for AWS operations
+AWS_PROFILE=
 
+# base image on top of which the actual Kubeflow image is built
+# (to speed up rebuilds)
 COMMON_IMAGE=
-IMAGE=
 
-OWNER=${user}
-DATASET_OWNER=${user}
-
-PROJECT=${project}
-
-TRAIN_TASK_NAME=
-
+# default version of the Genie components to include in the docker image
 genienlp_version=master
 genie_version=master
 thingtalk_version=master
+bootleg_version=master
