--- kubeflow/.gitignore
@@ -1,3 +1,0 @@
-!components/*.yaml
-.ipynb_checkpoints
-__pycache__
