+++ sync-repos.sh
